# purge archived only brings back archived records