Based upon feedback from http://talk.manageiq.org/t/vm-destroy-is-slow/2141/4

It was pointed out that we were bringing back every vm to determine if it was archived. This is slow.
Since that script was written, `archived` has become a virtual attribute and a scope. So we have faster means of accomplishing this request.

- use a scope and only bring back the archived vms.
- remove `REPORT_ONLY` which before turned the script into a no-op by default. Now it actually purges them by default.

Not sure if the `updated_on` was put into the query for performance or business reasons. I left it in there. Suggestions for us to remove this clause? It seems arbitrary.

| |before|after|
|---|---|---
|non-delete|30 seconds|3 seconds|