@chessbyte 2017-03-03
@kbrock I thought the `REPORT_ONLY` was acting like a Linux standard `--dry_run` option.  Maybe it makes sense to have that option available.  /cc @Fryguy