@Fryguy 2017-03-03
@chessbyte You are correct.  REPORT_ONLY is more or less "--dry-run" which I would like to keep for purging scripts as it allows a customer to run it and see what *might* change before deleting anything.  In purge_metrics (which is the most fleshed out of the scripts), we call this parameter `--mode`, and the default there is `count`.  The user has to opt-in to purging by changing that to `purge`