@Fryguy 2017-03-03
> Not sure if the updated_on was put into the query for performance or business reasons

If you remove the clause then all archived VMs will go away which is not the intent.  A user wants to know about recently archived VMs, but may want to clear out ones older than some time period (the ARCHIVE_CUTOFF)