@Fryguy 2017-03-03
> We should probably update all of the purge scripts at some point to a) have a common trollop based interface and b) use --dry-run instead of --mode as it's more intuitive

In hindsight, I change my mind.  --dry-run as an option would imply that the default action is to delete everything, which I don't want.  These scripts are *highly destructive* and so their default mode of operation should be non-destructive and just informative.  The user should be forced to opt-in via a parameter.  I'm open to better parameter names.