@kbrock 2017-03-04
@Fryguy Well, since the script is named `purge_archived_vms.rb` I had thought it would be destructive.

ok, anyway. sorry for deleting the stuff I thought was useful. I guess I brought upon the scope creep upon myself