@kbrock 2017-03-06
@Fryguy ok, rolled back all the code.

I just made the change to not bring back every vm and instead only bring back vms to purge.