@kbrock 2017-03-10
NOTE: if this gets backported, ensure that `Vm.all_archived` no longer has the `to_a` - if it does, then make fix accordingly.